Add delete button for Mathjax
